cuckoo: Add Cuckoo REST API job expiry
Our current bookkeeping of running Cuckoo jobs did not take into
account that jobs might never finish for whatever reason. This would
leave us accumulating more and more samples in the running jobs list
(think memory leak) and poll endlessly an ever increasing number of
jobs.

This change adds job expiry based on age to the running job list. To
keep the necessary data nicely bundled together we add a class CuckooJob
that uses its creation time as submission time and can be interrogated
as to its age. Otherwise it just stores the sample as before.

Add some wrapper methods to the Cuckoo class to wrap all accesses to the
running jobs list and necessary locking as well as the CuckooJob
abstraction.

Since this expiry might leave client starving, we also need a mechanism
to tell them about our abandoning the job. To achive this we add the
concept of Cuckoo analysis failure to Sample. If we drop a job because
it became too old, we mark the analysis as failed in the sample and
resubmit it to the work queue. In the ruleset we check this failure
marker and return the same error value as for submit errors in the first
place, leading to the same error message which incidentally already
reads that analysis has failed. :)

As a side-effect this finally allows us to fail analysis if we're unable
to get the report of a finished analysis for whatever reason.

Closes #137.