amavis: Add content disposition to meta info
The content disposition of the original MIME part is interesting for
deciding whether to ignore a sample or not because e.g. inline parts
would likely never be offered to the user for opening with another
application but rendered by the mail client instead.

Add extraction of the content disposition to our amavis plugin by monkey
patching our way around some amavis internals. It's not pretty but
works.

Register the new kind of meta info with Sample and add a property to
access is. Extend the testsuite to check for regressions. Add an example
rule to ruleset.conf.sample.