toolbox: Introduce Filetools and FiletoolsReport
Filetools is an analyser that can be used to create a report on the file and its
contents.

It specifically produces a file type from either file content or file name.

Currently the sample object determines MIME type for a given sample in a number
of ways: a declared type (optionally) communicated to us by the client, types
determined by the magic module based on file content and types determined by
the magic module based on file extension. All of them are put into a single set
to be used by our white- and greylisting rules. As such, the property can also
be used in expression rules.

There's two ways in which this can be improved though:

1. There are cases where it would beneficial if declared type and determined
types could be addressed individually, e.g. a rule that compares the declared
with determined types and triggers further analysis on discrepancies.

2. The sample actively determining stuff about itself makes it an analyser. We
already have a concept of analysers and implement it for oletools and cuckoo.
It would be straightforward to have a file property analyser as well.

This change introduces an explicit Filetools analyser which produces a
FiletoolsReport using the existing mime type determination functions. This
allows to address the discrete elements individually for more fine-grained
expressions.

Michael Weiser: Extended commit message and minor adjustments to the
code to address pylint suggestions.