expressions: Implement operators and, or and xor
The infixNotation parser of our expressions module already accepted the
logical operators and (&), or (|) and xor (^) but failed to actually
implement them. This is easily fixed by adding the necessary handling
logic to the EvalArith class.

In the context of this changeset these operators are useful and used to
combine sets (set()|set()) through their overloaded implementation for
them.

Michael Weiser: Extended commit message.