filetools: Move guess_ functions inside Filetools class
Make guess_mime_type_from_file_contents and
guess_mime_type_from_filename methods of the Filetools analyser to keep
functionality bundled where it belongs and make it clear that Filetools
are now in charge of this kind of analysis. There should be (and now can
be) no direct uses of these functions any more. There's actually no
common infrastructure in Filetools for them to benefit from yet, which
pylint correctly complains about. But we're expecting to extend the
class to provide added benefit and common code to those methods in the
future.

Michael Weiser: Extended commit comment and minor adjustments to the
code to address pylint suggestions.