test: Add rtf and rtx file extensions to expressions
RTX is an alternative extension to RTF. Both are sometimes used by
malware to mask that they're actually a DOC containing macros. So we add
the extensions to the sample expression in ruleset.conf.sample and
extend the testsuite to check for regressions on this.

Michael Weiser: Extended commit message and added the extensions to
ruleset.conf.sample.