oletools: Switch Oletools analyser to be per-sample
Currently Oletools has no state or infrastructure that needs it to be
long-lived. Indeed we throw it away after each analysis. Therefore we
can simplify its methods and use by making it a per-sample object. The
sample it is to work on is given to it upon initialisation and from then
on its methods work on that sample. With this we can drop method
arguments and move registration of the report with the sample into the
analyser which is a nice similarity to the Cuckoo analyser. The caching
decision, i.e. whether to use an already present report from the sample,
also moves there.

Address some linter warnings while touching this code.