filetools: Restore usage of declared name
Our removal of Sample.mime_types also removed the usage of the declared
name if available. This name is very useful because it is not only the
original intended file name but is also much more likely to include a
file extension that may give a hint to the magic/mimetype modules what
they're dealing with. Add code to Filetools that goes looking for that
declared name if helpful to the analysis.