expressions: Check identifier and attribute existance
If an expression accessed an unknown identifier or nonexistant
attribute we would only detect that at runtime and throw an ugly
backtrace in the log. Extend ExpressionsRule to run every expression
through a context of empty dummy objects on startup to detect these
kinds of errors. Print a concise error message to that effect and bail.

Unfortunately we have to pull all the report classes into rules.py to
instantiate the dummies there. At least we try to keep the interface
minimal and allow reports to be empty. This incidentally highlighted
some assumptions inside the report classes about their initialisers
which were no longer true and needed hardening against.

Closes #141.