docs: Sync install instructions to reality
Update installation instructions, particularly in main README to match
current best practice. Consistently use python interpreter binary names
including versions but without paths for clarity and shorter syntax as
well as portability.

Also, hopefully, finally get the package to install to get the
virtualenv tool right. The situation (on Ubuntu 18.04 at least) seems to
be:

- virtualenv pulls in python3-virtualenv and provides a virtualenv tool
  based on python3 that is also capable of creating python2 virtualenvs.
- python-virtualenv contains the python2 version of the module but still
  pulls in package virtualenv for the /usr/bin/virtualenv tool which is
  hardcoded to python3 and therefore pulls in python3-virtualenv

So if all we want is a working /usr/bin/virtualenv we can just install
the virtualenv package and not care about the python-version-specific
odule packages at all.