server: Make server the main loop
Previously, handling Cuckoo output in embed mode would define our
daemon lifetime. With embed mode gone it becomes clear that tracking
Cuckoo jobs is a backend concern and our main focus is servicing client
requests. Therefore we make the server the daemon's main loop. This
allows for client requests to define daemon lifetime, e.g. implement a
shutdown request efficiently.

For this we turn the Cuckoo module into a "job tracker" running in its
own thread. This tracker thread is requested in truly non-blocking
fashion to shut down e.g. by the signal handler or main program and a
new close_down() method eventually waits for the thread to exit.

The Peekaboo server is still running in its own thread because that is
required for socketserver's shutdown() method to work. But our
PeekabooServer wrapper class now waits for being requested to shut down
through the non-blocking shut_down() interface used by the signal
handler and then triggers a synchronous shutdown. The same interface can
potentially be used by other modules or individual requests to trigger a
shutdown.

As a side-effect, shutdown through signals is effectively parallelized
by registering the server as well as the job tracker with the signal
handler so both immediately start shutting down immediately upon
signals.

The server module is also reorganised to more clearly reflect
responsibilities: The PeekabooStreamServer is reduced to additional
functiuonality necessary for tracking requests and providing them with
infrastructure. Securing and removing of the connection socket is moved
to the PeekabooServer wrapper class where it fits more naturally and
together with gradual booting of the server through
bind_and_activate=False provides for better error handling.