queueing: Further speed up shutdown
Make the queue's shut_down() method non-blocking so it can be used by
the signal handler. Registering the queue with the signal handler fully
parallelises shutdown (server, job tracker and queue are notified in
parallel). Because we only trigger worker shutdown and then turn to
something else for a short while, actual closing down, i.e. waiting for
the workers to end, avoids one sleep when idle because the workers are
usually shut down by then.

While at it, eliminate the useless running_flag and use is_alive() of
the thread object instead to determine if the worker is still alive or
shut down.