cuckoo: Eliminate submission via sample
We have been on a path to remove analyser submission from the Sample.
While it seems like a nice object-oriented concept to say that the sample
submits itself to the analyser, it's not actually the sample driving
this but the ruleset and the benefit of doing it via the sample is very
low (saving the job ID for diagnostics and some logging to the internal
report). The sample basically just served as transport vehicle of the
analyser.

Moving submission into the ruleset removes one layer of indirection and
clarifies sample flow and component interaction.

We could just pass the Cuckoo analyser / job tracker to the ruleset and
be done with it. But since we anticipate to add more analysers that need
long-lived job trackers, we take this opportunity to implement this more
generically: A new analyzer configuration file and object is added which
contains analyzer configuration. Only this configuration is passed to
the ruleset engine. The engine creates the job trackers and hands them
to the rules if they need them. For this each rule is marked by a class
variable or property to need a particular job tracker.

As a nice side-effect, this puts the ruleset engine into a position to
only start job trackers that are actually needed by rules. Expression
mechanics are extended to allow querying whether an expression contains
a particular identifier (independent of whether evaluation will actually
dereference it) so that e.g. the Cuckoo job tracker can be startet if an
expression may request it's report.

Since job trackers may need a reference to the job queue to submit
samples back into it after analysis, the ruleset engine is now created
by the job queue (again) and a reference to the job queue passed into it
for handing to the created job trackers to avoid a circular object
creation order dependency.