queueing: Separate object construction from init
In order to be able to shut down quickly during startup we must separate
object creation from initialisation so that the object constructs
quickly, can be registered with the signal handler and then be shut down
quickly during slower resource initialisation. This is particularly true
for the Cuckoo API connectivity test which is now somewhat tucked away
in daemon->queue->ruleset engine->cuckoo.

This also requires to check whether shutdown was requested during this
lengthy initialisation and shut down any resources which may have been
allocated, since they may not have existed yet when the request arrived.
This is employed in the ruleset engine to shut down job trackers.

Also, we add a flag to the signal handler so we can abort startup if
we're asked to shut down again to avoid getting stuck in a
half-inited/half-shut-down state. This problem likely existed all the
time and we just never ran into it.