cuckoo: Handle various invalid JSON API responses
This change changes the handling of invalid or incomplete Cuckoo API
responses to be more comprehensive, robust and easier to understand.
In particular it adds checks for the response value being None or not of
type int.

It also removes logging of the original JSON in error/warning message so
we do not pollute them with content which isn't under our control. If
this is ever needed again, it should be added as separate debug logging.

Closes #159.