server: Raise receive buffer size limit
When the client tries to hand us a large number of samples for analysis,
the JSON will at some point be more than 16KiB, depending mainly on the
length of the filenames used. We return an invalid JSON error in that
case which the client could use to chunk its request to stay below the
limit.

We've hit this problem with our AMaViS plugin and since we don't want to
complicate it with chunking its requests when we want to switch to a
completely different protocol anyway, we just raise the receive buffer
maximum to 512KiB. This should be large enough for any (sensible,
non-DoS) real-world emails and still prevent us becoming DoS-able.

Closes #163.