toolbox: Introduce Knowntools and KnowntoolsReport
Knowntools is an analyser that can be used to check the database if this sample
has already been analysed.

As of now it only has one property `known` which can be True or False, and
`result` which holds the cached result.

Before expression rules could not evaulate whether the sample is known or not.

Now expression rules can be constructed that check if a sample is known and
ignore only with certain constraints. One usecase could be to only trust cached
`bad` results and continue analysis of Sample is unknown or not detected
malicious.

This analyser will be extended to support more in the future.