db: Remove table AnalysisJournal
We don't use this table for anything so I remove it.

The AnalysisJournal table is not used by any software component of peekaboo.
It has advantages for analysts to have this information but also results in this
extra table, foreign key relations, and potentially results in expensive queries
with joins.

A limitation of this separate AnalysisJournal is the lack of the actual result
of this individual analysis.
Which is why in future likely a timestamp will be added to SampleInfo which
then allows multiple entries for the same Sample and results in an easier and
more flexible journal there.

For now this leads to creating an entry in SampleInfo with every analysis_save
that can only be distinguished from the other entries by its id.