sample: Constrain file extension character set
Restrict the character set allowed for file extensions to avoid
submitting bogus extensions to analyzers. We've seen Cuckoo being
sensitive to non-ASCII- and whitespace characters in filenames. We still
keep the set of allowed punctuation characters wide open to catch
exploitation attempts based on those. What we won't be able to catch any
more will be usage of very similar Unicode glyphs to confuse the user
and trick them into opening something that looks like a .DOCX (but is
actually something like a .D0CX) with Word, bypassing virus scanners.
But since Cuckoo is not able to handle these anyway, that's no loss.
Needs to be revisited once we start using analysers which are able to
handle these.

Closes #165.