db: Retry deadlocked transactions
Due to some peculiar locking behaviour, in MySQL/MariaDB even the
simplest INSERT and DELETE operations/transactions can deadlock on a
standalone server. This is detected by the server though and one of the
competing transactions aborted. We get an OperationalError exception on
this and need to retry the transaction.

Add the necessary helper routine and handling in our in-flight-locking
routines, where we've been hit by this.

Closes #164.