db: Retry on additional failure scenarios
There are a number of additional failure scenarios for database queries
we regularly encounter. One is the server closing the connection, e.g.
because it is restarting. The other is the server not being reachable
(yet) because it is still restarting. This change adds handling for both
via the connection_invalidated property of SQLAlchemy statement-related
exceptions and adding a back-off retry for the MySQL-specific connection
failed error code. This makes aggressive connection pool recycling
unnecessary, so we can remove it.

Closes #171.