oletools: Fix autoexec and suspicious extraction
What seems to be a change in behaviour of oletools or python in version
3.8 has showcased the fact that previously we were only adding the first
autoexec or suspicious finding from oletools to our analyzer report.
This made us blind to all others. Change the logic to extract all
findings.

This also makes the respective expression rule test work again which is
triggering on the literal VBOX.  Also add a new test to the oletools
tests to hopefully catch regressions in the future. Extend the test
logic to use regexes to be able to concisely formulate the expected
result of the test.