cortex: Selectively extract VirusTotal taxonomies
Don't just take the first element of the list of taxonomy results but
look for the first matching with namespace VT.

The rationale for the change is input validation: We do not
want to just accept anything but explicitly look for the expected
namespace and make sure there's only one entry.

A new helper function enforces a constraint
that there is only ever one taxonomy of namespace VT.

Because the overall JSON structure of reports in Cortex is similar
the new function get_element_from_list_of_dicts is added to the base
class and meant to handle the recuring pattern of lists of dicts found
in those reports.