cuckoo: Add input validation to report
Our Cuckoo report is very light on input validation and inefficient by
storing a full copy of the report dict returned by Cuckoo.

Explicitly extract supported properties from the Cuckoo report dict and
verify their types for robustness and security. Throw exceptions upon
violations and handle those exceptions when creating the report by
flagging the analysis as failed so the client is notified. Do not store
a full copy of the Cuckoo dict for memory efficiency. Add unit tests to
catch regressions.

Because we no longer store the full report, we can not dump it from
Sample.dump_processing_report() upon failure. Instead, replace property
raw of the report with one called dump which produces a defined
structure similar to the original report, containing all the information
we have.