rest api: Switch the server to sanic
Use the sanic framework to switch the server from a UNIX domain socket
to a REST API. This allows for remote communication as well as
structured input and output. The API has two endpoints for now:

/v1/scan - accepts a sample for analysis as form-data file upload in
field name 'file'

/v1/report/<job_id> - returns a (minimal for now) job report for the
supplied job ID

This conversion fundamentally changes the operation mode and parameters
of Peekaboo. To list a few:

We no longer get handed a list of file names of Amavis temporary MIME
part files. Instead the file is uploaded via HTTP. We do not write it
out to a temporary file but keep it in memory (but could "swap it out"
based on a threshold to conserve memory if we wanted). Therefore it is
handed to our analysers (Ole, file, Cuckoo) by content not file name any
more. This change makes most of the legacy logic in Sample superfluous
so it can be removed.

We no longer allow the client to wait for the analysis to finish on the
scan endpoint. Instead it has to poll the report endpoint for the
report.

The client therefore has to do more work. In the case of our AMaViS
plugin it now needs to submit the individual samples, track their
analysis state and construct an overall verdict from the individual
reports.

The file metadata is transferred as standard form field properties
Content-Type and filename (instead of a magic meta_info structure).
Content-Disposition could betransferred as a form field property as well
but unfortunately sanic does not allow access to that property (while
aiohttp and flask would). So we transfer it as a global HTTP request
header for now. The actual HTTP Content-Disposition header has a
different meaning than that of MIME multipart email messages. So we add
an X- prefix to its name.

The frontend server is single threaded for now which presents a
significant bottleneck, particularly if database operations go into
retries. The plan is to switch more and more of the input processing and
database access to asyncio in later commits to mitigate that effect.

When we used to initialise the sample, we also added an entry to the
report stating that analysis had begun. This could now go into the
constructor. Apart from immediately triggering generation of the
sha256sum it is also simply superfluous now since we no longer do an
initialisation of the sample. So just remove it for good.