rest api: Add textual verdict to job report
Add a textual representation of the analysis result to the job report to
be more client-friendly. Employ it in peekaboo-util. This should get us
one step closer to a consistent client interface on the report endpoint.

Make reason a separate field in anticipation of the actual report
becoming available in the report field as soon as we find a way to
preserve it in the database.

Eliminate the rating from the REST API return structure because it
exposes our internal result enum values. They are carefully chosen to
allow numeric comparison of 'bad-' and 'failed-ness' as an
implementation detail. Making them part of the REST API would make it
impossible to adjust them if requirements change. Clients should instead
rely on the textual representation of the result in the 'result' field
and possibly map those to their own numeric classification system for
easy comparsion.