cortex: Rename tlp parameters to make pylint happy
Rename the argument to not mask the global enum and mark additional
kwargs to the method in CortexHashAnalyzer as unused as suggested by
pylint. We choose the del statement because fiddling with **_ throws up
problems when using a parameter with default value without explicit
mention of its name. Incidentally, we turn the method into a class
method as we don't actually use self.