db: Ignore unfinished analyses for the journal
Interrupting Peekaboo while an analysis is in progress may leave it in
an unfinished state in the database. This would lead to another analysis
of an identical sample to pick up that entry and report the sample as
analysed with result 'unchecked'. This would pollute the sample info
cache with invalid data and prevent future analyses from producing a
correct result.

Filter for finished analyses when retrieving the analysis journal from
the database.