rest api: Sync up the testsuite
Adjust the testsuite to all the invasive changes to Sample in particular
to make it work again.

Most changes stem from the eleminitation of the SampleFactory. Instead
we can now create samples directly in memory with defined content
without creating temporary files.

Another case is where we want to use defined content from test data
files. This is implemented using a new subclass of Sample called
FileSample which reads a given file into itself and initialises its
filename as well if not overridden.

Finally there are cases where we just need a Sample with a given
filename but without actual content. Here we intentionally set the
content to None so exceptions are raised if accesses to it are tried.

Remove a duplicate import of module schema while at it.