db: Silence aiosqlite
The aiosqlite module picks up the global log level unconditionally.
Therefore any raising of the Peekaboo log level will switch on
additional messages of aiosqlite as well. We don't want that. Instead we
set its log level to our separate DB log level explicitly to silence it
(for most of the time).