amavis: Streamline plugin
Now that the REST API is somewhat stabilised we can streamline the
AMaViS plugin somewhat. Jobs are now submitted directly from the
metadata extraction loop and only the job ID is rememebered for tracking
purposes.

The tracking loop exits as soon as a failed or bad result is encountered
or a communication error with Peekaboo occurs. This is based on the fact
that these kinds of errors are definite and cannot get better again.
Therefore we can report a final failure or virus found verdict to
AMaViS and forget about the other jobs. Otherwise we eventually return
the classic 'No malware found' at the very end.

The result text is accumulated across both loops which nicely ends up
resembling a kind of virus scanner output to be presented to the admin
for debugging and possibly even the user.

The tracking loop sleeps for a polling interval between loops. For the
very first iteration we scale that polling interval based on the number
of jobs submitted and a very small factor to quickly catch jobs which
were answered quickly based on cached results.

For ease of configuration we add a default for the API base URL. A new
parameter is added to adjust the polling interval. The documentation is
adjusted to reflect these facts.