db: Allow for a list of asyncio drivers to try
Because asyncmy is very new and aiomysql supports older python versions,
we want to try both with a preference for asyncmy. Implement a logic
that allows for this and incidentally no longer needs regular
expressions.

Unfortunately this cements the fact that we only support selected
database backends and those only with selected asyncio drivers.