db: Allow overriding the asyncio database driver
Add a config option that can be used to override the asyncio database
driver and disable auto-detection. This is meant as a last resort for
cases where we don't know the database backend at all or are unaware of
an available asyncio driver for a backend we know in principle. Update
the test suite and add some documentation.