peekaboo-util: Fix common options
The previous attempt at getting common options to work before and after
the subcommand name failed in that it made them work only after it. The
reason lies in the peculiar way argparse works. So we go about it much
more explicitly and aggressively to get rid of this annoyance.