db: Eliminate asyncio lock
It turns out that only sqlite responds poorly to multiple modify
operations running on the same database through multiple connections
because the special default NullPool used for sqlite will spin up lots
of connections in multiple threads very quickly on high scan request
load. Switching to a QueuePool as used by default by all other backends
does not fully solve that problem. But customising that QueuePool to
provide only a single connection so that all modifies are serialised
does and allows us to get rid of the asyncio lock altogether.

Something similar might be possible with threaded accesses as well.