sample: Reduce the representation to bare minimum
The sample repr is unused currently. But in case we start using it
again, reduce it to the bare minimum that makes a sample identifyable.
This is its ID for now since we have only very small windows of time
where a sample may have no ID. Then we will represent it as
Sample(id='None').