Add asyncio.get_running_loop() for Python 3.6
Python 3.6 does not have asyncio.get_running_loop(). At the time,
calling asyncio.get_event_loop() was the only way to get at the running
event loop. Unfortunately this has the side-effect of silently starting
a new loop if there is none. We want to know if that were to happen,
either through an exception or follow-up errors. So we intentionally use
asyncio.get_running_loop. Now we monkey-patch it into asyncio if not
present based on the code of Python 3.8.