cuckoo, cortex: Avoid KeyError on job journal corruption
We want to raise an exception if we detect that Cuckoo or Cortex gave us
a job ID for a job that is already used by another job we are currently
tracking. For diagnostics we want to format both IDs of our own jobs
(samples) into the exception. The ID for the colliding old sample was
previously not retrieved correctly. A default string representation of
our CuckooJob or CortexJob objects was formatted instead. This was now
uncovered by our switch to numeric job IDs upon which the formatting no
longer works and raises an AttributeError (due to switching to the
Sample.id property as well). Fix this by retrieving the sample from the
CuckooJob or CortexJob and accessing its id attribute as intended.