cuckoo, cortex: Forget colliding jobs for correctness
If Cuckoo or Cortex give us the same job ID twice, we have already
successfully submitted the new job.  Therefore the old job can no longer
exist from the point of view of Cortex or Cuckoo. That would mean,
they've forgotten about it, for example because their databases have
been cleared. Therefore our tracking of the old job and failing the new
job is wrong because the analyzer results would be for a different (the
newer) sample.

So we turn the logic aound, fail the old job and start tracking the new
job as normal.