filetools: Handle empty filename
Previously each sample always had at least a filename generated by
AMaViS (p00x) and optionally an additional declared name. With our
switch to the REST API the former disappeared, Sample.filename and
Sample.name_declared are now aliases of each other and can be None. This
would lead to a NoneType exception in guess_mime_type_from_name() when
trying to fall back to Sample.filename from Sample.name_declared.

We change the logic to not fall back to Sample.filename and abort
MIME type guessing from the filename if we don't have one.

This eliminates the last internal user of Sample.name_declared. The only
reason for it to exist is compatibility with expression rules where it's
documented part of the available set of properties.

Note: This problem was masked because currently it's actually impossible
to submit a file without a filename through the API (about to be
changed).